Initial capability register support for core files
Capability registers are stored in a new note type NT_CAPREGS.  The
implementation is overall straightforward, although they are rendered
as printable strings in a rather hackish way right now: the CHERI
capability register format is hard-coded in the value rendering function
DataExtractor::Dump.

In a later change I would like to supply an arch-specific AST RecordDecl
to represent the capability register format and then use the struct
pretty-printer to render them.